- nearly completed switch from uint16_t to struct for glove-values, feature extractor changes in progress

+++ TangoHapps/classes/objects/conditions/THGestureClassifier.h
@@ -11,5 +11,5 @@
 @interface THGestureClassifier : TFSimulableObject
 @property (nonatomic, assign, readwrite) int numberOfTicksToDetect;
 @property (nonatomic, assign, readwrite) NSUInteger halfWindowSize;
-- (void)addSignal:(float)signal;
+- (void)addSignal:(uint32_t)signal;
 @end
